Complete new API call
The new API call allows for the buyer to see if an ask transaction has completed so they can know to delete it.

Also introduce a concept of the seller having seen the transaction was completed so they can request their money then the buyer can see it is completed.